% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{add_on_list}
\alias{add_on_list}
\title{add_on_list}
\usage{
add_on_list(add_on_list)
}
\arguments{
\item{add_on_list}{A list of ggplot2-add-ons that are supposed to be integrated in
the visualization process.}
}
\description{
add_on_list
}
