% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addClusterVariable}
\alias{addClusterVariable}
\alias{addGroupingVariable}
\title{Add discrete/categorical variables}
\usage{
addClusterVariable(
  object,
  cluster_df,
  cluster_name,
  phase = "first_tmt",
  overwrite = FALSE,
  verbose = TRUE
)

addGroupingVariable(object, input_df, variable_name, phase = "first_tmt")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{cluster_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\code{cluster_name}}{Character or factor. The cluster variable that is to be joined.}
}}

\item{cluster_name, variable_name}{Character value. The name of the variable that is to be joined.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}

\item{input_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\code{variable__name}}{Character or factor. The discrete variable that is to be joined.}
}}
}
\value{
An updated cell tracer object that contains the data added.
}
\description{
These functions allow to join new discrete/categorical variables that can be referred
to via the \code{across}-argument of many functions.
}
