% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addDimRedData}
\alias{addDimRedData}
\title{Add dimensional reduction data}
\usage{
addDimRedData(object, dim_red_df, dim_red_method)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{dim_red_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\emph{x}}{Numeric. The cell id's position on the x-axis.}
\item{\emph{y}}{Numeric. The cell id's position on the y-axis.}
}}

\item{dim_red_method}{Character value. The dimensional reduction method.}
}
\value{
An updated cell tracer object that contains the data added.
}
\description{
This function allows to add dimensional reduction data to the
object that can be referred to e.g. in \code{plotDimRed()}.
}
