% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-default.R
\name{adjustDefaultInstructions}
\alias{adjustDefaultInstructions}
\title{Adjust object based default}
\usage{
adjustDefaultInstructions(object, ...)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{...}{The argument names with their new default value specified with the
following syntax 'argument_name' = 'argument_default'.

Use \code{adjustableDefaultInstructions()} to obtain all argument names
whose default input you can manipulate this way.}
}
\description{
The default input for many recurring arguments is carried with
the celltracer object. Use this function to savely adjust the default
according to your preference.
}
\examples{

 object <- adjustDefaultInstructions(object, pt_clrp = "jco", pt_alpha = 0.6, method_dist = "spearman")


}
