% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{aes_to}
\alias{aes_to}
\title{aes_to}
\usage{
aes_to(color_to, fill_to, shape_to)
}
\arguments{
\item{color_to}{Character value. The variable that is to be displayed or highlighted by the dots color.}

\item{fill_to}{Character value. The varaible that is to be displayed or highlighted by the dots filling.}

\item{shape_to}{Character value. The variable that is to be displayed or highlighted by the dots shape.}
}
\description{
aes_to
}
