% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{agglomerateHierarchicalCluster}
\alias{agglomerateHierarchicalCluster}
\title{Title}
\usage{
agglomerateHierarchicalCluster(
  object,
  phase = NULL,
  method_dist = NULL,
  method_aggl = NULL,
  force = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dist}{}
}
\value{

}
\description{
Title
}
