% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-family.R
\name{animateAllTracks}
\alias{animateAllTracks}
\title{Title}
\usage{
animateAllTracks(
  object,
  across = "condition",
  across_subset = NULL,
  frame_subset = NULL,
  tmt = "afterwards",
  n_cells = 100,
  color_to = NULL,
  linetype = "solid",
  linesize = 0.75,
  add_on_list = list(),
  ...
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed. Valid input
options can be obtained via \code{getAcrossOptions()}.}

\item{across_subset}{Character vector. The groups of interest that the \code{across}-
variable contains. Valid input options can be obtained via \code{getVariableValues()}.}

\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}

\item{linetype}{Character value. Valid options are \emph{'solid', 'twodash', 'longdash', 'dotted'}
and \emph{'dotdash'}.}

\item{linesize}{Numeric value. Denotes the size of the lines drawn.}

\item{add_on_list}{A list of ggplot2-add-ons that are supposed to be integrated in
the visualization process.}

\item{...}{Additional arguments given to \code{gganimte::animate()}.}
}
\value{
An animated GIF.
}
\description{
Title
}
