% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-family.R
\name{animateWellImage}
\alias{animateWellImage}
\title{Animate cellular movement}
\usage{
animateWellImage(
  object,
  well_plate,
  well,
  image,
  style = "normal",
  add_on_list = list(),
  ...
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{well}{Character value. The well of interest (e.g. \emph{'A1'}, \emph{'B12'})}

\item{image}{Numeric value. The well-image of interest.}

\item{style}{Character value. If set to \emph{'cell_tracker'} imitates the tracking process of the CellTracker software.}

\item{add_on_list}{}

\item{...}{Additional arguments given to \code{gganimate::animate()}.}
}
\value{
An animated GIF.
}
\description{
This function animates the movement of cells that were covered
within one well-image.
}
