% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{argument_dummy}
\usage{
argument_dummy(pt_alpha, pt_clr, pt_clrp, pt_clrsp, pt_size)
}
\arguments{
\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{discrete_feature}{Character value. Specifies the name of the grouping variable
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{scales, ncol, nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
argument_dummy
}
