% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_track_list_shiny}
\alias{assemble_track_list_shiny}
\title{Integrate cell track tables}
\usage{
assemble_track_list_shiny(track_data_list, well_plate_list, object)
}
\arguments{
\item{track_data_list}{The output of \code{load_cell_track_files_shiny}.}
}
\value{
A track data.frame
}
\description{
Subsets the successfully loaded files
}
