% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-default.R
\name{assign_default}
\alias{assign_default}
\title{Assigns default to parent function}
\usage{
assign_default(object)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\description{
Assigns default to parent function
}
