% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_object}
\alias{check_object}
\title{Check Celltracer Object Input}
\usage{
check_object(object, experiment = NULL, set_up_req = "process_data")
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\description{
Makes sure that object input is of class celltracer and
that it contains the relevant information needed for the function.
}
