% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{check_pam_input}
\alias{check_pam_input}
\title{Check pam input}
\usage{
check_pam_input(k)
}
\arguments{
\item{k}{NUmeric value. The k-value used to calculate the pam-cluster.}
}
\description{
Check pam input
}
