% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_phase}
\alias{check_phase}
\title{Check phase input}
\usage{
check_phase(object, phase, max_phases = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{A character value referring to the phase.}

\item{max_phase}{Numeric value or NULL. If numeric it regulates the maximal number of phases allowed.}
}
\description{
Check phase input
}
