% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{check_smooth}
\alias{check_smooth}
\title{Check smooth}
\usage{
check_smooth(smooth, smooth_se, smooth_span)
}
\arguments{
\item{smooth}{Logical. If set to TRUE the values are smoothed.}

\item{smooth_se}{Logical. If set to TRUE the standard error will be displayed.}

\item{smooth_span}{NUmeric value. Denotes the smoothing span used.}
}
\description{
Check smooth
}
