% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_wp_directories}
\alias{check_wp_directories}
\title{Detect double directories}
\usage{
check_wp_directories(all_wp_lists)
}
\arguments{
\item{all_wp_lists}{A list of well plate lists.}
}
\value{
A character value that contains either \emph{'unique'} or the
directories that are not unique.
}
\description{
Detect double directories
}
