% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-initiation.R
\name{checkDataQuality}
\alias{checkDataQuality}
\title{Object initiation: Step 3}
\usage{
checkDataQuality(object)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\value{
An updated celltracer object.
}
\description{
Opens an interactive application in which the quality of the data can be assessed and
where the data set can be filtered accordingly.
}
