% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{cluster_df}
\alias{cluster_df}
\title{cluster_df}
\usage{
cluster_df(cluster_df)
}
\arguments{
\item{cluster_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\code{cluster_name}}{Character or factor. The cluster variable that is to be joined.}
}}
}
\description{
cluster_df
}
