% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{colors}
\alias{colors}
\title{colors}
\usage{
colors(clrp, clrsp)
}
\arguments{
\item{clrp}{Character value. The color panel to be used. Valid input options can be obtained
via \code{allColorpanels()}.}

\item{clrsp}{Character value. The color spectrum to be used for continuous variables. Valid
input options can be obtained via \code{allColorspectra()}.}
}
\description{
colors
}
