% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cto.R, R/initiate-cto.R
\name{compileCto}
\alias{compileCto}
\title{Compiles a cell tracer object}
\usage{
compileCto(ct_list, in_shiny = FALSE, compute_umap = FALSE)

compileCto(ct_list, in_shiny = FALSE, compute_umap = FALSE)
}
\arguments{
\item{ct_list}{}
}
\value{
A cell tracer object.
}
\description{
Compiles a cell tracer object

Compile the cell tracer object
}
