% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-kmeans.R
\name{computeKmeansCluster}
\alias{computeKmeansCluster}
\title{Title}
\usage{
computeKmeansCluster(
  object,
  phase = NULL,
  method_kmeans = NULL,
  k = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_kmeans}{}

\item{k}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
