% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{computePamCluster}
\alias{computePamCluster}
\title{Title}
\usage{
computePamCluster(
  object,
  phase = NULL,
  method_pam = NULL,
  k = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_pam}{}

\item{k}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
