% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-family.R
\name{computeUmapData}
\alias{computeUmapData}
\title{Compute Umap data}
\usage{
computeUmapData(object)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
An updated cell tracer object.
}
\description{
A wrapper around the umap algorithm and \code{addDimRedData()}.
}
