% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{correlateAll}
\alias{correlateAll}
\alias{correlateAcross}
\title{Title}
\usage{
correlateAll(
  object,
  phase = NULL,
  method_corr = NULL,
  print_errors = FALSE,
  verbose = NULL
)

correlateAcross(
  object,
  phase = NULL,
  across = NULL,
  method_corr = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_corr}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
