% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{detect_outliers_iqr}
\alias{detect_outliers_iqr}
\title{Outlier detection functions}
\usage{
detect_outliers_iqr(object, verbose = NULL, variable_names = NULL)
}
\description{
Outlier detection functions
}
