% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Detect outliers}
\usage{
detectOutliers(object, method_outlier = "iqr", verbose = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated celltracer object.
}
\description{
There are several methods according to which outliers can be detected or
defined. Which one to use strongly depends on the researchers question at hand. See details
for more information on which method does what exactly.
}
\details{
This function only detects outliers. It does not remove them from the object. Use \code{removeOutliers()}
for that matter.
}
\seealso{
removeOutliers()
}
