% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{dim_red_method}
\alias{dim_red_method}
\title{dim_red_method}
\usage{
dim_red_method(dim_red_method)
}
\arguments{
\item{dim_red_method}{Character value. The dimensional reduction method.}
}
\description{
dim_red_method
}
