% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{discardDistanceMatrix}
\alias{discardDistanceMatrix}
\title{Title}
\usage{
discardDistanceMatrix(object, phase = NULL, method_dist = NULL, verbose = NULL)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dist}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
