% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doPamClustering}
\alias{doPamClustering}
\title{Title}
\usage{
doPamClustering(
  object,
  k_values,
  stand = TRUE,
  phase = "first_tmt",
  variables = "all",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{k_values}{Numeric vector. Denotes all k-values of interest.}

\item{stand}{Given to argument \code{stand} of function \code{cluster::pam()}. Denotes
whether the numeric variables are scaled before the clsutering process or not. Default
is TRUE.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}

\item{variables}{The numeric variables upon which the clusters are calculated. Valid inputs can
be obtained with the function \code{getStatVariableNames()}.

The default is set to \emph{'all'} which includes all numeric values.}

\item{...}{Additional arguments given to \code{cluster::pam()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated celltracer-object.
}
\description{
Performs pam-clustering with the respective variables, during
the denoted phase for all values denoted in input for argument \code{k_values}.

The name of the respective clustering results is assembled as a combination of
\emph{'pam_k'} and the respective value for k.
}
\seealso{
getPamObject(), getClusterNames()
}
