% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests.R
\name{existOutlierResults}
\alias{existOutlierResults}
\title{Logical Tests}
\usage{
existOutlierResults(object)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\value{

}
\description{
Logical tests that return TRUE if the specified functions/algorithms
have been used on the object and FALSE if not.
}
