% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCategoricalVariablesNames}
\alias{getCategoricalVariablesNames}
\title{Helper functions to extract}
\usage{
getCategoricalVariablesNames(object, phase = NULL)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Helper functions to extract
}
