% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCellIds}
\alias{getCellIds}
\title{Get cell ids}
\usage{
getCellIds(object, na_max = 0, ...)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{na_max}{Numeric value. Sets the threshold of the maximum number of missing values
a cell can have across all variables. Ignored if set to NULL or if the experiment
is not of type \emph{timelapse}.}
}
\value{
Character vector of cell ids.
}
\description{
Allows to filter cell ids according to a variety of aspects.
}
