% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCellLines}
\alias{getCellLines}
\alias{getConditions}
\title{Obtain cell line and condition names}
\usage{
getCellLines(object)

getConditions(object, phase = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\value{
Character vector.
}
\description{
Quick wrapper around the functionality of getGroupingVariableNames(). As
the cell line does not change from experiment phase to experiment phase there is no
need to specify it.
}
\details{
Useful helper function when it comes to specify conditions and cell lines
of interest via the \code{across_subset}-argument.
}
