% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getClusterData}
\alias{getClusterData}
\title{Obtain cluster data}
\usage{
getClusterData(object, phase = "first_tmt")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{
A data.frame that contains the cell ids and their cluster belonging.
}
\description{
Obtain cluster data
}
