% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getClusterNames}
\alias{getClusterNames}
\title{Obtain valid cluster names}
\usage{
getClusterNames(object, phase = "first_tmt", algorithm_subset = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}

\item{algorithm_subset}{Character value or NULL. If character only the cluster
names of the algorithms denoted are returned.}
}
\value{
A list named according to all cluster algorithms found. Each named slot
contains a character vector of all cluster names of the algorithm.
}
\description{
Returns all valid cluster names that can be used as input
for the \code{across}-argument.
}
