% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getData}
\alias{getData}
\title{Obtain data slots}
\usage{
getData(object, data_slot, phase)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{data_slot}{Character value. One of \emph{'stats', 'tracks', 'meta'} or \emph{'cluster'}.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{
The data.frame of interest.
}
\description{
A wrapper around \code{purr::map_df()} and the respective
list of the data slot of interest.
}
