% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDimRed}
\alias{getDimRed}
\title{Obtain dimensional reduction data}
\usage{
getDimRed(
  object,
  dim_red_method = "umap",
  with_meta = TRUE,
  with_stats = TRUE,
  phase = "first_tmt"
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{dim_red_method}{Character value. The dimensional reduction method.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_stats}{Logical. If set to TRUE the numeric variables of the stat data slots are added
tot he output data.frame.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{
The data.frame of interest.
}
\description{
Obtain dimensional reduction data
}
