% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingDf}
\alias{getGroupingDf}
\title{Obtain cluster data}
\usage{
getGroupingDf(object, phase = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame that contains the cell ids and their cluster belonging.
}
\description{
Obtain cluster data
}
