% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\alias{getAcrossOptions}
\title{Obtain variable names that group the cells}
\usage{
getGroupingOptions(object, phase = NULL)

getAcrossOptions(object, phase = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
An informative list.
}
\description{
This function returns the names of the variables that
group cell ids and can therefore be used as input for the \code{across}
argument.
}
