% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupNames}
\alias{getGroupNames}
\alias{getGroups}
\title{Obtain group names a grouping variable contains}
\usage{
getGroupNames(object, option, phase = "all")

getGroups(object, option)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{option}{Character value. Denotes the discrete variable - the grouping of cells -
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}
}
\value{
Character vector of group names.
}
\description{
Obtain group names a grouping variable contains
}
