% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{getHclustObj}
\alias{getHclustObj}
\title{Title}
\usage{
getHclustObj(object, phase = NULL, method_dist = NULL, method_aggl = NULL)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dist}{}

\item{method_aggl}{}
}
\value{

}
\description{
Title
}
