% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMeta}
\alias{getMeta}
\title{Obtain meta data}
\usage{
getMeta(object)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
The data.frame of interest.
}
\description{
Obtain meta data
}
