% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMetaDf}
\alias{getMetaDf}
\title{Obtain meta data}
\usage{
getMetaDf(object, phase = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\value{
A data.frame with all variables providing meta information about cells of
the track data.frame. (such as condition, cell line, well plate belonging etc.)
}
\description{
Obtain meta data
}
