% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPamObject}
\alias{getPamObject}
\title{Obtain pam-clustering results.}
\usage{
getPamObject(object, k, phase = "first_tmt")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{k}{Numeric value. The k-value for which the pam-object of interest has been computed.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{
A pam object - the return value of the function \code{cluster::pam()} that has been
computed via the function \code{doPamClustering()}.
}
\description{
Obtain pam-clustering results.
}
