% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{getPcaConv}
\alias{getPcaConv}
\alias{getTsneConv}
\alias{getUmapConv}
\title{Obtain dimensional reduction objects}
\usage{
getPcaConv(object, phase = NULL)

getTsneConv(object, phase = NULL)

getUmapConv(object, phase = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An S4 object of class \emph{'dim_red_conv'}.
}
\description{
Returns the S4 objects in which the dimensional reduction
results are stored.
}
