% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getStatsDf}
\alias{getStatsDf}
\alias{getStats}
\title{Obtain stat data.frame}
\usage{
getStatsDf(object, phase = NULL, with_cluster = NULL, with_meta = NULL)

getStats(object, phase = NULL, with_cluster = NULL, with_meta = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
A data.frame with all numeric variables summarizing the measurements of
the track data.frame.
}
\description{
Obtain stat data.frame
}
