% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getStatVariableNames}
\alias{getStatVariableNames}
\title{Obtain all numeric stat-variables}
\usage{
getStatVariableNames(object)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
A character vector.
}
\description{
Obtain all numeric stat-variables
}
