% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTracksDf}
\alias{getTracksDf}
\alias{getTracks}
\title{Obtain track data.frame.}
\usage{
getTracksDf(
  object,
  phase = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  verbose = NULL
)

getTracks(
  object,
  phase = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
A data.frame in which each observation refers to a cell at a given frame.
}
\description{
Obtain track data.frame.
}
