% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateDf}
\alias{getWellPlateDf}
\title{Obtain well plate information}
\usage{
getWellPlateDf(object, well_plate = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{well_plate}{Character value. Denotes the well plate of interest. Use function \code{getWellPlateNames()}
to obtain all valid input options.}
}
\value{

}
\description{
Obtain well plate information
}
