% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_caption_add_on}
\alias{hlpr_caption_add_on}
\title{Clarifying caption}
\usage{
hlpr_caption_add_on(object, phase)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
\item{\emph{'entire'}}{Uses the complete data of the entire timespan.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{

}
\description{
Returns a \code{ggplot2::labs()}-add on in which the
caption clarifies whether the plot refers to the times pan before the treatment,
after the treatment or to the entire experiment.

If there was not treatment or the cells were treated right from the beginning
no caption is returned.
}
