% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_process_tracks}
\alias{hlpr_process_tracks}
\title{Processing helper}
\usage{
hlpr_process_tracks(df, phase, object, verbose)
}
\description{
To be used as mapped function in purrr::map_*(). Only used
if experiment type is 'time_lapse'. Processes the track data.frame
with regards to column names as well as some migration data computation if
x- and y-coords are available.
}
