% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{initiateCorrelation}
\alias{initiateCorrelation}
\title{Title}
\usage{
initiateCorrelation(
  object,
  phase = NULL,
  variables_subset = NULL,
  force = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{force}{}

\item{verbose}{}

\item{...}{}
}
\value{

}
\description{
Title
}
