% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate-cto.R
\name{initiateCTO}
\alias{initiateCTO}
\title{Create a Cell Tracer Object}
\usage{
initiateCTO()
}
\value{
A cell tracer object.
}
\description{
This function allows for interactive creation of a
cell-tracer-object.
}
