% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{initiateHierarchicalClustering}
\alias{initiateHierarchicalClustering}
\title{Title}
\usage{
initiateHierarchicalClustering(
  object,
  phase = NULL,
  variables_subset = NULL,
  force = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{force}{}

\item{verbose}{}

\item{...}{}
}
\value{

}
\description{
Title
}
