% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{initiatePamClustering}
\alias{initiatePamClustering}
\title{Title}
\usage{
initiatePamClustering(
  object,
  phase = NULL,
  variables_subset = NULL,
  force = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{variables_subset}{}

\item{force}{}

\item{verbose}{}

\item{...}{}
}
\value{

}
\description{
Title
}
