% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-celltracer.R
\name{launchCellTracer}
\alias{launchCellTracer}
\title{Access interactive application}
\usage{
launchCellTracer(object = NULL)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\description{
Opens an interactive application in which most of the
data visualization can be done conveniently.
}
