% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{linetype}
\alias{linetype}
\title{linetype}
\usage{
linetype(linetype)
}
\arguments{
\item{linetype}{Character value. Valid options are \emph{'solid', 'twodash', 'longdash', 'dotted'}
and \emph{'dotdash'}.}
}
\description{
linetype
}
