% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{n_cells}
\alias{n_cells}
\title{n_cells}
\usage{
n_cells(n_cells)
}
\arguments{
\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}
}
\description{
n_cells
}
