% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{phase_cluster}
\alias{phase_cluster}
\title{phase_cluster}
\usage{
phase_cluster(phase)
}
\arguments{
\item{phase_cluster}{Character value. Refers to the phase of the experiment from which to take
the clustering variables. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\description{
phase_cluster
}
