% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{plotAvgSilhouetteWidths}
\alias{plotAvgSilhouetteWidths}
\title{Title}
\usage{
plotAvgSilhouetteWidths(
  object,
  k = NULL,
  phase = NULL,
  clr = "steelblue",
  display_cols = TRUE,
  display_line = TRUE,
  display_points = TRUE
)
}
\arguments{
\item{object}{}

\item{k}{}

\item{phase}{}

\item{clr}{}

\item{display_cols}{}

\item{display_line}{}

\item{display_points}{}
}
\value{

}
\description{
Title
}
