% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotCellCount}
\alias{plotCellCount}
\title{Plot cell count}
\usage{
plotCellCount(object, across, color_by)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed. Valid input
options can be obtained via \code{getAcrossOptions()}.}
}
\value{
A ggplot.
}
\description{
Visualizes the number and distribution of cells across
a discrete feature of choice.
}
