% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{plotCorrelationSD}
\alias{plotCorrelationSD}
\title{Plot correlation variance across groups}
\usage{
plotCorrelationSD(
  object,
  method_corr = NULL,
  phase = NULL,
  across = NULL,
  relevel = NULL,
  variables_subset = NULL,
  signif_level = NULL,
  clrsp = NULL
)
}
\arguments{
\item{across}{Character vector (!). Denotes all grouping variables of interest.}
}
\value{


A ggplot.
}
\description{
Uses the style of correlation matrices to visualize
the standard deviation of the correlation values across groups of
a grouping variable - indicating which grouping is responsible for changes
in correlation.
}
