% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{plotCorrplot}
\alias{plotCorrplot}
\title{Title}
\usage{
plotCorrplot(
  object,
  phase = NULL,
  method_corr = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  variables_subset = NULL,
  plot_type = "lower",
  display_diagonal = TRUE,
  signif_level = NULL,
  clr_low = "darkred",
  clr_high = "steelblue",
  shape = "tile",
  shape_size = 15,
  size_aes = FALSE,
  display_values = TRUE,
  values_alpha = 0.9,
  values_clr = "black",
  values_digits = 2,
  values_size = 4,
  draw_grid = TRUE,
  grid_clr = "black",
  grid_size = 0.5,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_corr}{}

\item{across}{}

\item{across_subset}{}

\item{relevel}{}

\item{variables_subset}{}

\item{plot_type}{}

\item{display_diagonal}{}

\item{signif_level}{}

\item{clr_low}{}

\item{clr_high}{}

\item{shape}{}

\item{shape_size}{}

\item{size_aes}{}

\item{display_values}{}

\item{values_alpha}{}

\item{values_clr}{}

\item{values_digits}{}

\item{values_size}{}

\item{draw_grid}{}

\item{grid_clr}{}

\item{grid_size}{}

\item{ncol}{}

\item{nrow}{}
}
\value{

}
\description{
Title
}
