% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Title}
\usage{
plotDendrogram(
  object,
  phase = NULL,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  branch_size = 1,
  display_legend = TRUE,
  display_title = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dist}{}

\item{method_aggl}{}

\item{k}{}

\item{h}{}

\item{branch_size}{}

\item{display_legend}{}

\item{display_title}{}

\item{clrp}{}

\item{clrp_adjust}{}

\item{ncol}{}

\item{nrow}{}

\item{...}{}
}
\value{

}
\description{
Title
}
