% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot dimensional reduction}
\usage{
plotDimRed(
  object,
  dim_red = "umap",
  color_to = NULL,
  pt_size = 1,
  pt_alpha = 0.9
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{color_to}{Character value. The variable that is to be displayed or highlighted by the dots color.}

\item{pt_size}{Numeric value. Denotes the size and alpha values of the points of the scatterplot.}

\item{pt_alpha}{Numeric value. Denotes the size and alpha values of the points of the scatterplot.}
}
\value{
A ggplot.
}
\description{
Visualizes the dimensional reduction method of choice.
}
