% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotPamClusterCount}
\alias{plotPamClusterCount}
\title{Plot Number of Cells in Cluster}
\usage{
plotPamClusterCount(
  object,
  k,
  phase = "first_tmt",
  clrp = "milo",
  flip_coords = FALSE,
  ...
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{k}{NUmeric value. The k-value used to calculate the pam-cluster.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}

\item{clrp}{Character value. The color panel to be used. Valid input options can be obtained
via \code{allColorpanels()}.}

\item{flip_coords}{Logical. Reverses x- and y-axis}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}
}
\value{
A ggplot.
}
\description{
Visualizes the number of cells in each cluster.
}
