% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{plotPamMedoids}
\alias{plotPamMedoids}
\title{Title}
\usage{
plotPamMedoids(
  object,
  phase = NULL,
  k = NULL,
  method_pam = NULL,
  clrp = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{k}{}

\item{method_pam}{}

\item{clrp}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
