% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotPamSilWidth}
\alias{plotPamSilWidth}
\title{Plot Cluster Quality}
\usage{
plotPamSilWidth(
  object,
  k,
  phase = "first_tmt",
  clrp = "milo",
  display_average = TRUE
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{k}{NUmeric value. The k-value used to calculate the pam-cluster.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}

\item{clrp}{Character value. The color panel to be used. Valid input options can be obtained
via \code{allColorpanels()}.}

\item{display_average}{Logical. If set to TRUE the average across all clusters
is displayed via a horizontal, dashed line.}
}
\value{
A ggplot.
}
\description{
Visualizes the cluster quality via the average silhouette Width.
}
