% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{plotPca}
\alias{plotPca}
\alias{plotTsne}
\alias{plotUmap}
\title{Plot dimensional reduction results}
\usage{
plotPca(
  object,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  clrp = NULL,
  clrp_adjust = NULL,
  clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL
)

plotTsne(
  object,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  clrp = NULL,
  clrp_adjust = NULL,
  clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL
)

plotUmap(
  object,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  clrp = NULL,
  clrp_adjust = NULL,
  clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{color_by}{Character value or NULL. If character, denotes either the numeric- or grouping variable whoose values
are displayed by color. If set to NULL the color is specified by the argument \code{pt_clr}.}

\item{color_aes}{Character value. Only relevant if \code{color_by} is specified.
Denotes the aesthetic with which colors are displayed. Either \emph{'color'} or \emph{'fill'}. Depending on that as well as the input for argument
\code{pt_shape} the design of the geometric objects (points) varies according to the rules
of the ggplot2-framework.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}
}
\value{


A ggplot.
}
\description{
Visualizes dimensional reduction in a scatterplot.
}
