% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-kmeans.R
\name{plotScreeplot}
\alias{plotScreeplot}
\title{Title}
\usage{
plotScreeplot(
  object,
  phase = NULL,
  method_kmeans = NULL,
  k = NULL,
  clr = "steelblue",
  display_cols = TRUE,
  display_line = TRUE,
  display_points = TRUE
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_kmeans}{}

\item{k}{}

\item{display_cols}{}

\item{display_line}{}

\item{display_points}{}
}
\value{

}
\description{
Title
}
