% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{plotSilhouetteWidths}
\alias{plotSilhouetteWidths}
\title{Title}
\usage{
plotSilhouetteWidths(
  object,
  k = NULL,
  method_pam = NULL,
  clrp = NULL,
  ncol = NULL,
  nrow = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{k}{}

\item{method_pam}{}

\item{clrp}{}

\item{ncol}{}

\item{nrow}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
