% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotSingleTracks}
\alias{plotSingleTracks}
\title{Plot single cell migration}
\usage{
plotSingleTracks(object, cell_ids, color_by = NULL, scales = "free")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{cell_ids}{Character vector. Denotes the cell ids of interest.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}
}
\value{
A ggplot.
}
\description{
Visualizes the migration of single cells of interest.
}
