% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotWellPlate}
\alias{plotWellPlate}
\title{Plot the well plate set up}
\usage{
plotWellPlate(
  object,
  well_plate = "",
  color_to = NULL,
  fill_to = "cl_condition"
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{color_to}{Character value. The variable that is to be displayed or highlighted by the dots color.}

\item{fill_to}{Character value. The varaible that is to be displayed or highlighted by the dots filling.}
}
\value{
A ggplot.
}
\description{
Plot the well plate set up
}
