% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{pretty_names}
\alias{pretty_names}
\title{pretty_names}
\usage{
pretty_names(pretty_names)
}
\arguments{
\item{pretty_names}{Logical. If set to TRUE the function attempts to convert the concisely named
variables into more aesthetically pleasing ones.}
}
\description{
pretty_names
}
