% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{pt_args}
\alias{pt_args}
\title{pt_args}
\usage{
pt_args(pt_color, pt_size, pt_alpha)
}
\arguments{
\item{pt_color}{Character value. Denotes the color of the points of the scatterplot.}

\item{pt_size, pt_alpha}{Numeric value. Denotes the size and alpha values of the points of the scatterplot.}
}
\description{
pt_args
}
