% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-content.R
\name{renameCluster}
\alias{renameCluster}
\title{Rename cluster of cells}
\usage{
renameCluster(object, phase = NULL, cluster_variable = NULL, ...)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{cluster_variable}{Character value. The name of the cluster variable
whoose cluster are to be renamed.}

\item{...}{The clusters to be renamed specified according to the following
syntax: \emph{'new_cluster_name'} \code{=} \emph{'old_cluster_name'}.}
}
\value{
An updated celltracer object.
}
\description{
Allows to rename particular clusters within a cluster variable
of the celltracer object's overall data.
}
