% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helper.R
\name{rotateLabelsX}
\alias{rotateLabelsX}
\alias{rotateLabelsY}
\title{Rotate labels of ggplot output}
\usage{
rotateLabelsX(angle = 90)

rotateLabelsY(angle = 90)
}
\arguments{
\item{angle}{Numeric value. Denotes the angle with which
the labels are to be rotated.}
}
\value{

}
\description{
Allows to rotate the labels of the x-/y-axis
of ggplot2 plots. Useful in case of overlap.
}
