% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{run_dim_red}
\alias{run_dim_red}
\title{Title}
\usage{
run_dim_red(
  object,
  phase = NULL,
  variables_subset = NULL,
  method_dim_red = "pca",
  force = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{variables_subset}{}

\item{force}{}

\item{verbose}{}

\item{...}{}
}
\value{

}
\description{
Title
}
