% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setClusterConv}
\alias{setClusterConv}
\alias{setDimRedConv}
\title{Set clustering object}
\usage{
setClusterConv(object, cluster_object, method, phase, variable_set)

setDimRedConv(object, dim_red_object, method, phase, variable_set)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}
}
\description{
Sets a convenient object and empties its data slot.
}
