% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setStorageDirectory}
\alias{setStorageDirectory}
\title{Set default storage directory}
\usage{
setStorageDirectory(object, directory_cto)
}
\arguments{
\item{object}{A valid cypro object.}

\item{directory_cto}{Character value. The directory under which
the object is supposed to be stored via \code{saveCelltracerObject()}.}
}
\description{
Set default storage directory
}
