% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-celltracer-object.R
\name{subsetByCellId}
\alias{subsetByCellId}
\alias{subsetByCellLine}
\alias{subsetByCondition}
\alias{subsetByCluster}
\title{Create subsets of your experiment}
\usage{
subsetByCellId(object, cell_ids, phase = "all", verbose = NULL, ...)

subsetByCellLine(object, cell_lines, phase = "all", verbose = NULL)

subsetByCondition(object, conditions, phase = NULL, verbose = NULL)

subsetByCluster(
  object,
  cluster_variable,
  cluster,
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid celltracer object.}

\item{cell_ids}{Character vector. Denotes the cells to keep unambiguously with their cell ids.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{cell_lines}{Character vector. Denotes the cell lines to be kept.}

\item{conditions}{Character vector. Denotes the conditions to be kept.}

\item{cluster_variable}{Character value. Denotes the clustering variable to be used in order to subset
the celltracer object.}

\item{cluster}{Character vector. Denotes the cluster of the clustering variable specified in argument
\code{cluster_variable} to be kept.}
}
\value{
A celltracer object that contains the data for the subsetted cells over the specified phase(s).
}
\description{
These function allow to conveniently split your data by certain characteristics such
as cell lines, conditions, cluster etc. This might be useful if you want apply some machine learning
techniques such as clustering and correlation on only a subset of cells.
}
\details{
The valid input options for argument \code{phase} vary As clustering is computed for every
phase respectively and a cell's condition might differ from phase to phase the functions \code{subsetByCondition()} and
\code{subsetByCluster()} need a character value as input for argument \code{phase} which filters the
data by the respective input as well as by the specified phase. Functions \code{subsetByCellId()} and \code{subsetByCellLine()} allow to create subsets over more than
just one phase.

Correlation, clustering and dimensional reduction results must be computed again as the data set changes.
}
