% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-celltracer-object.R
\name{updateCelltracerObject}
\alias{updateCelltracerObject}
\title{Keep celltracer object up to data}
\usage{
updateCelltracerObject(object)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\value{
An updated celltracer object.
}
\description{
Ensures that the celltracer object is constructed according
to the latest version of the package.
}
