% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{verbose}
\alias{verbose}
\title{variables_num}
\usage{
verbose(verbose)
}
\arguments{
\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{variables}{Character vector. Denotes the numeric variables of interest.
Run \code{getNumericVariableNames()} with you celltracer-object to obtain
all valid input options.
verbose}
}
\description{
variables_num
}
