% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{well_plate}
\alias{well_plate}
\title{well_plate}
\usage{
well_plate(well_plate)
}
\arguments{
\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\description{
well_plate
}
