% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{with_cluster}
\alias{with_cluster}
\title{with_clsuter}
\usage{
with_cluster(with_cluster)
}
\arguments{
\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}
}
\description{
with_clsuter
}
