% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{with_meta}
\alias{with_meta}
\title{with_meta}
\usage{
with_meta(with_meta)
}
\arguments{
\item{with_meta}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}
}
\description{
with_meta
}
