% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{with_stats}
\alias{with_stats}
\title{with_stats}
\usage{
with_stats(with_stats)
}
\arguments{
\item{with_stats}{Logical. If set to TRUE the numeric variables of the stat data slots are added
tot he output data.frame.}
}
\description{
with_stats
}
