% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addClusterVariable}
\alias{addClusterVariable}
\title{Add discrete/categorical variables}
\usage{
addClusterVariable(
  object,
  cluster_df,
  cluster_name,
  phase = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{cluster_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\code{cluster_name}}{Character or factor. The cluster variable that is to be joined.}
}}

\item{cluster_name, variable_name}{Character value. The name of the variable that is to be joined.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated cell tracer object that contains the data added.
}
\description{
Allows to join new discrete/categorical variables that can be referred
to via the \code{across}-argument of many functions.
}
