% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{addHierarchicalCluster}
\alias{addHierarchicalCluster}
\title{Title}
\usage{
addHierarchicalCluster(
  object,
  phase = NULL,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dist}{}

\item{method_aggl}{}

\item{k}{}

\item{h}{}
}
\value{

}
\description{
Title
}
