% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addPamCluster}
\alias{addPamCluster}
\title{Add PAM clustering results to overall data}
\usage{
addPamCluster(
  object,
  phase = NULL,
  method_pam = NULL,
  k = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_pam}{Character vector (or value see details for more.) Denotes the algorithm of interest.
Valid input options are \emph{'euclidean'} and \emph{'manhattan'}.}

\item{k}{Numeric vector. All k-values of interest.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated celltracer object.
}
\description{
Adds the clustering results of \code{computePamClusters()} in form
of grouping variables to the object's overall data - making them available for the \code{across}-
argument.
}
\details{
The last step of the PAM clustering pipeline. This function iterates
over all combinations of \code{method_pam} and \code{k} and
adds the respective clustering variable to the object's overall data named according to
the following syntax: \emph{pam_\code{method_pam}_k_\code{k}}.

Use \code{getgroupingOptions()} afterwards to obtain the exact names.
}
