% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCellLines}
\alias{getCellLines}
\alias{getConditions}
\title{Obtain cell line and condition names}
\usage{
getCellLines(object)

getConditions(object, phase = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
Character vector.
}
\description{
Obtain cell line and condition names
}
\details{
Useful helper function when it comes to specify conditions and cell lines
of interest via the \code{across_subset}-argument.
}
