% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\alias{getAcrossOptions}
\title{Obtain variable names that group the cells}
\usage{
getGroupingOptions(object, phase = "first_tmt")

getAcrossOptions(object, phase = "first_tmt")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. Refers to the phase of the experiment. Valid inputs are:

\describe{
\item{\emph{'before_tmt'}}{Uses the data that refers to the time before the treatment.}
\item{\emph{'first_tmt'}}{Uses the data that refers tot the time after the treatment.}
}

Note: This argument is ignored if the experiment set up did not include any treatment or the treatment
started right from the beginning.}
}
\value{
An informative list.
}
\description{
This function returns the names of the variables that
group cell ids and can therefore be used as input for the \code{across}
argument.
}
