% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMetaDf}
\alias{getMetaDf}
\title{Obtain meta data}
\usage{
getMetaDf(object, phase = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}
}
\value{
The data.frame of interest.
}
\description{
Obtain meta data
}
