% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateDf}
\alias{getWellPlateDf}
\title{Obtain well plate information}
\usage{
getWellPlateDf(object, well_plate = NULL)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A data.frame in which each row contains information about a well.
}
\description{
Obtain well plate information
}
