% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotCellCount}
\alias{plotCellCount}
\title{Plot cell count}
\usage{
plotCellCount(object, across, color_by)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
cells of your experiment in a certain manner.}
}
\value{
A ggplot.
}
\description{
Visualizes the number and distribution of cells across
a discrete feature of choice.
}
