% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotSingleTracks}
\alias{plotSingleTracks}
\title{Plot single cell migration}
\usage{
plotSingleTracks(object, cell_ids, color_to = NULL, scales = "free")
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{cell_ids}{Character vector. Denotes the cell ids of interest.}

\item{color_to}{Character value. The variable that is to be displayed or highlighted by the dots color.}

\item{scales}{Character value. Given to argument \code{scales} of
\code{ggplot2::facet_wrap()}.}
}
\value{
A ggplot.
}
\description{
Visualizes the migration of single cells of interest.
}
