% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotWellPlate}
\alias{plotWellPlate}
\title{Plot the well plate set up}
\usage{
plotWellPlate(
  object,
  well_plate = NULL,
  color_by = NULL,
  clrp_adjust = NULL,
  make_pretty = NULL
)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A ggplot.
}
\description{
Plot the well plate set up
}
