% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-content.R
\name{renameCluster}
\alias{renameCluster}
\title{Rename cluster of cells}
\usage{
renameCluster(object, phase = NULL, cluster_variable = NULL, ...)
}
\arguments{
\item{object}{A valid cell tracer object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'} or \emph{'second'}).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{cluster_variable}{Character value. The name of the cluster variable
whoose cluster are to be renamed.}

\item{...}{The groups to be renamed specified according to the following
syntax: \emph{'new_group_name'} \code{=} \emph{'old_group_name'}.}
}
\value{
An updated cto object.
}
\description{
Allows to rename groups within a cluster variable
of the cluster data of the specified phase.
}
