% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{saveCelltracerObject}
\alias{saveCelltracerObject}
\title{Save Celltracer Object}
\usage{
saveCelltracerObject(object, verbose = TRUE)
}
\arguments{
\item{object}{A valid celltracer object.}
}
\description{
Save Celltracer Object
}
