% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{saveCelltracerObject}
\alias{saveCelltracerObject}
\title{Save cypro object}
\usage{
saveCelltracerObject(object, verbose = TRUE)
}
\arguments{
\item{object}{A valid cypro object.}
}
\description{
Save cypro object
}
